# OpenApi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Fitzo API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Users/profile": {
      "post": {
        "tags": [
          "Users"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserProfileDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserProfileDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Users/bmr": {
      "get": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "formula",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/BmrFormula"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BmrFormula": {
        "enum": [
          "MifflinStJeor",
          "HarrisBenedict"
        ],
        "type": "string"
      },
...
      "Gender": {
        "enum": [
          "Male",
          "Female"
        ],
        "type": "string"
      },
...
      "UserProfileDto": {
        "type": "object",
        "properties": {
          "weight": {
            "type": "number",
            "format": "double"
          },
          "height": {
            "type": "number",
            "format": "double"
          },
          "age": {
            "type": "integer",
            "format": "int32"
          },
          "gender": {
            "$ref": "#/components/schemas/Gender"
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Please enter a valid token",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}